+++ streamsx/.toolkit/com.ibm.streamsx.topology/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changes
 ==========
 
+## v1.15.5
+* [#2499](https://github.com/IBMStreams/streamsx.topology/issues/2499) Pythondoc: Updated `Stream.parallel`,`Stream.set_parallel` args description regarding submission parameter 
+* [#2498](https://github.com/IBMStreams/streamsx.topology/issues/2498) Python: Support SPL window creation using submission parameter 
+* [#2494](https://github.com/IBMStreams/streamsx.topology/issues/2494) JAVA: Support window creation using submission parameter 
+* [#2495](https://github.com/IBMStreams/streamsx.topology/issues/2495) edgeConfig: `condaPackages` parameter is not supported anymore
+
 ## v1.15.4
 * [#2488](https://github.com/IBMStreams/streamsx.topology/issues/2488) execution context: Make submission time parameters available without serializing the parameter callable